Refactor and implementation of firebase firestore database

+++ app/src/main/java/Database/SQLLite/BudgetDBHandler.kt
@@ -1,7 +1,7 @@
-package Database
+package Database.SQLLite
 
+import CustomErrors.CustomException
 import Models.Budget
-import Models.Category
 import android.annotation.SuppressLint
 import android.content.ContentValues
 import android.content.Context
@@ -14,7 +14,8 @@ class BudgetDBHandler(context: Context, name: String?, factory: SQLiteDatabase.C
     private val TABLE_BUDGET = "BUDGETS"
     private val TABLE_CATEGORY = "CATEGORIES"
 
-    private val CREATE_BUDGET = ("CREATE TABLE IF NOT EXISTS " + TABLE_BUDGET
+    private val CREATE_BUDGET = (
+            "CREATE TABLE IF NOT EXISTS " + TABLE_BUDGET
             + "(_ID INTEGER PRIMARY KEY AUTOINCREMENT, "
             + "TITLE TEXT UNIQUE NOT NULL, "
             + "DESCRIPTION TEXT,"
@@ -40,8 +41,9 @@ class BudgetDBHandler(context: Context, name: String?, factory: SQLiteDatabase.C
     }
 
     override fun onUpgrade(db: SQLiteDatabase, oldVersion: Int, newVersion: Int) {
-
-        db.execSQL("DROP TABLE IF EXISTS $TABLE_BUDGET")
+        if(newVersion > 1){
+            db.execSQL("DROP TABLE IF EXISTS $TABLE_BUDGET")
+        }
 
         onCreate(db)
     }
@@ -50,14 +52,18 @@ class BudgetDBHandler(context: Context, name: String?, factory: SQLiteDatabase.C
         val values = ContentValues().apply {
             put(COLUMN_TITLE, budget.title)
             put(COLUMN_DESC, budget.description)
-            put(COLUMN_BUDGET, budget.budget)
-            put(COLUMN_CATEGORY_ID, budget.category_id)
+            put(COLUMN_BUDGET, budget.value)
+            put(COLUMN_CATEGORY_ID, budget.categoryId)
         }
 
         val db = this.writableDatabase
 
-        db.insert(TABLE_BUDGET, null, values)
-        db.close()
+        try{
+            db.insert(TABLE_BUDGET, null, values)
+            db.close()
+        }catch (e: SQLiteException){
+            throw CustomException("Error on item creation: ${e.message}")
+        }
     }
 
     @SuppressLint("Range")
@@ -73,25 +79,31 @@ class BudgetDBHandler(context: Context, name: String?, factory: SQLiteDatabase.C
             return ArrayList()
         }
 
-        var id: Int
+        //var id: Int
+        var id : String
         var title: String
         var description: String
-        var budget: Double
-        var category_id: Int
-
-
-        if (cursor!!.moveToFirst()) {
-            while (!cursor.isAfterLast) {
-                id = cursor.getString(cursor.getColumnIndex(COLUMN_ID)).toInt()
-                title = cursor.getString(cursor.getColumnIndex(COLUMN_TITLE))
-                description = cursor.getString(cursor.getColumnIndex(COLUMN_DESC))
-                budget = cursor.getString(cursor.getColumnIndex(COLUMN_BUDGET)).toDouble()
-                category_id = cursor.getString(cursor.getColumnIndex(COLUMN_CATEGORY_ID)).toInt()
-
-                budgets.add(Budget(title, description, budget, category_id, id))
-                cursor.moveToNext()
+        var value: Double
+        //var category_id: Int
+        var categoryId: String
+        try{
+            if (cursor!!.moveToFirst()) {
+                while (!cursor.isAfterLast) {
+                    //id = cursor.getString(cursor.getColumnIndex(COLUMN_ID)).toInt()
+                    id = cursor.getString(cursor.getColumnIndex(COLUMN_ID))
+                    title = cursor.getString(cursor.getColumnIndex(COLUMN_TITLE))
+                    description = cursor.getString(cursor.getColumnIndex(COLUMN_DESC))
+                    value = cursor.getString(cursor.getColumnIndex(COLUMN_BUDGET)).toDouble()
+                    //category_id = cursor.getString(cursor.getColumnIndex(COLUMN_CATEGORY_ID)).toInt()
+                    categoryId = cursor.getString(cursor.getColumnIndex(COLUMN_CATEGORY_ID))
+                    budgets.add(Budget(id, title, description, value, categoryId))
+                    cursor.moveToNext()
+                }
             }
+        }catch (e: SQLiteException){
+            throw CustomException("Error on item reading: ${e.message}")
         }
+
         return budgets
     }
 
@@ -108,24 +120,32 @@ class BudgetDBHandler(context: Context, name: String?, factory: SQLiteDatabase.C
             return ArrayList()
         }
 
-        var id: Int
+        //var id: Int
+        var id: String
         var title: String
         var description: String
         var budget: Double
-        var category_id: Int
+        //var category_id: Int
+        var category_id: String
 
-        if (cursor!!.moveToFirst()) {
-            while (!cursor.isAfterLast) {
-                id = cursor.getString(cursor.getColumnIndex(COLUMN_ID)).toInt()
-                title = cursor.getString(cursor.getColumnIndex(COLUMN_TITLE))
-                description = cursor.getString(cursor.getColumnIndex(COLUMN_DESC))
-                budget = cursor.getString(cursor.getColumnIndex(COLUMN_BUDGET)).toDouble()
-                category_id = cursor.getString(cursor.getColumnIndex(COLUMN_CATEGORY_ID)).toInt()
-
-                budgets.add(Budget(title, description, budget, category_id, id))
-                cursor.moveToNext()
+        try{
+            if (cursor!!.moveToFirst()) {
+                while (!cursor.isAfterLast) {
+                    //id = cursor.getString(cursor.getColumnIndex(COLUMN_ID)).toInt()
+                    id = cursor.getString(cursor.getColumnIndex(COLUMN_ID))
+                    title = cursor.getString(cursor.getColumnIndex(COLUMN_TITLE))
+                    description = cursor.getString(cursor.getColumnIndex(COLUMN_DESC))
+                    budget = cursor.getString(cursor.getColumnIndex(COLUMN_BUDGET)).toDouble()
+                    //category_id = cursor.getString(cursor.getColumnIndex(COLUMN_CATEGORY_ID)).toInt()
+                    category_id = cursor.getString(cursor.getColumnIndex(COLUMN_CATEGORY_ID))
+                    budgets.add(Budget(id, title, description, budget, category_id))
+                    cursor.moveToNext()
+                }
             }
+        }catch (e: SQLiteException){
+            throw CustomException("Error on item reading: ${e.message}")
         }
+
         return budgets
     }
 
@@ -135,17 +155,22 @@ class BudgetDBHandler(context: Context, name: String?, factory: SQLiteDatabase.C
         val values = ContentValues().apply {
             put(COLUMN_TITLE, budget.title)
             put(COLUMN_DESC, budget.description)
-            put(COLUMN_BUDGET, budget.budget)
-            put(COLUMN_CATEGORY_ID, budget.category_id)
+            put(COLUMN_BUDGET, budget.value)
+            put(COLUMN_CATEGORY_ID, budget.categoryId)
         }
 
         val selection = COLUMN_ID + " LIKE ?"
         val selectionArgs = arrayOf(id)
-        val count = db.update(
-            TABLE_BUDGET,
-            values,
-            selection,
-            selectionArgs)
+
+        try{
+            val count = db.update(
+                TABLE_BUDGET,
+                values,
+                selection,
+                selectionArgs)
+        }catch (e: SQLiteException){
+            throw CustomException("Error on item update: ${e.message}")
+        }
     }
 
     fun delete(id: String){
@@ -155,7 +180,11 @@ class BudgetDBHandler(context: Context, name: String?, factory: SQLiteDatabase.C
 
         val selectionArgs = arrayOf(id)
 
-        db.delete(TABLE_BUDGET, selection, selectionArgs)
+        try{
+            db.delete(TABLE_BUDGET, selection, selectionArgs)
+        }catch (e: SQLiteException){
+            throw CustomException("Error on item creation: ${e.message}")
+        }
     }
 
     fun exists(categoryID: String) : Boolean{
@@ -186,22 +215,24 @@ class BudgetDBHandler(context: Context, name: String?, factory: SQLiteDatabase.C
             return ArrayList()
         }
 
-        var id: Int
+        //var id: Int
+        var id: String
         var title: String
         var description: String
         var budget: Double
-        var category_id: Int
-
+        //var category_id: Int
+        var category_id: String
 
         if (cursor!!.moveToFirst()) {
             while (!cursor.isAfterLast) {
-                id = cursor.getString(cursor.getColumnIndex(COLUMN_ID)).toInt()
+                //id = cursor.getString(cursor.getColumnIndex(COLUMN_ID)).toInt()
+                id = cursor.getString(cursor.getColumnIndex(COLUMN_ID))
                 title = cursor.getString(cursor.getColumnIndex(COLUMN_TITLE))
                 description = cursor.getString(cursor.getColumnIndex(COLUMN_DESC))
                 budget = cursor.getString(cursor.getColumnIndex(COLUMN_BUDGET)).toDouble()
-                category_id = cursor.getString(cursor.getColumnIndex(COLUMN_CATEGORY_ID)).toInt()
-
-                budgets.add(Budget(title, description, budget, category_id, id))
+                //category_id = cursor.getString(cursor.getColumnIndex(COLUMN_CATEGORY_ID)).toInt()
+                category_id = cursor.getString(cursor.getColumnIndex(COLUMN_CATEGORY_ID))
+                budgets.add(Budget(id, title, description, budget, category_id))
                 cursor.moveToNext()
             }
         }
